+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ### Fixes
-- Fix Crashes with Forge/NeoForge
+- Part 3 content registry, Screen Handler things!
+- Bump version